+++ packages/runtime/src/interpreter/opcode-list.ts
@@ -4961,3 +4961,89 @@ export class Base64Decode extends Op {
 		return this.computeCost();
 	}
 }
+
+export class Replace extends Op {
+	readonly line: number;
+	start: number;
+	original: Uint8Array;
+	replace: Uint8Array;
+
+	constructor(start: number, line: number) {
+		super();
+		this.line = line;
+		this.start = start;
+		this.original = new Uint8Array;
+		this.replace = new Uint8Array;
+	}
+
+	execute(stack: TEALStack): number {
+		if (this.start + this.replace.length > this.original.length) {
+			throw new RuntimeError(RUNTIME_ERRORS.TEAL.BYTES_REPLACE_ERROR, {
+				lenReplace: this.replace.length,
+				index: this.start,
+				lenOriginal: this.original.length,
+				line: this.line,
+			});
+		}
+		const result = new Uint8Array(this.original.length);
+		for (let i = 0, j = 0; i < this.original.length; ++i) {
+			if (i >= this.start && i < this.start + this.replace.length) {
+				result[i] = this.replace[j++];
+			}
+			else {
+				result[i] = this.original[i];
+			}
+		}
+		stack.push(result);
+		return this.computeCost();
+	}
+}
+
+/**
+ * Opcode: replace2 s
+ * Stack: ..., A: []byte, B: []byte → ..., []byte
+ * Copy of A with the bytes starting at S replaced by the bytes of B. Fails if S+len(B) exceeds len(A)
+ */
+export class Replace2 extends Replace {
+	/**
+	 * Asserts 0 arguments are passed.
+	 * @param args Expected arguments: [start_index]
+	 * @param line line number in TEAL file
+	 */
+	 constructor(args: string[], line: number) {
+		assertLen(args.length, 1, line);
+		super(Number(args[0]), line);
+	}
+
+	execute(stack: TEALStack): number {
+		this.assertMinStackLen(stack, 2, this.line);
+		this.replace = this.assertBytes(stack.pop(), this.line);
+		this.original = this.assertBytes(stack.pop(), this.line);
+		return super.execute(stack);
+	}
+}
+
+/**
+ * Opcode: replace3 
+ * Stack: ..., A: []byte, B: uint64, C: []byte → ..., []byte
+ * Copy of A with the bytes starting at B replaced by the bytes of C. Fails if B+len(C) exceeds len(A)
+ */
+ export class Replace3 extends Replace {
+	/**
+	 * Asserts 0 arguments are passed.
+	 * @param line line number in TEAL file
+	 */
+	 constructor(args: string[], line: number) {
+		assertLen(args.length, 0, line);
+		super(0, line);
+	}
+
+	execute(stack: TEALStack): number {
+		this.assertMinStackLen(stack, 3, this.line);
+		this.replace = this.assertBytes(stack.pop(), this.line);
+		this.start = Number(this.assertBigInt(stack.pop(), this.line));
+		this.original = this.assertBytes(stack.pop(), this.line);
+		return super.execute(stack);
+	}
+
+}